Fix strict compiler warning: cast float to double in printf to avoid -Wdouble-promotion error

+++ src/main.c
@@ -4,7 +4,7 @@
 
 int main() {
     float temp = get_temperature_celsius();
-    printf("Temperature: %.2f C\n", temp);
+    printf("Temperature: %.2f C\n", (double)temp);
     printf("Status: %s\n", check_temperature_status(temp));
     return 0;
 }
